Add sitting animation for Capybara entity
Three files in the CapyMod project have been adjusted to include sitting functionality for the Capybara entity. The CapybaraModel class now includes an additional check for sitting pose to trigger the animation. Animation states have been added to the CapybaraEntity class to manage the new action. Also, the Animation definition for 'SIT' has been specified in the ModAnimationDefinitions file.